Fixed all modules
Haha, 100th, lovely!

Modules are fixed

Beyond that,

fixed the issue where the logged-in user had an empty dropdown list display if they had no saved blocks yet

taken away some comments, console logs (incomplete cleanup)
