+++ src/blockchain/block.js
@@ -1,6 +1,6 @@
-import { calculateHash } from "../blockchain/calculateHash.js";
 
 export default class Block {
+
   constructor(data, newHash, previousHash) {
     this.data = data;
     this.previousHash = previousHash;
