+++ src/blockchain/calculateHash.js
@@ -21,11 +21,3 @@ export async function calculateHash(data) {
   return hashHex;
 }
 
-/* export async function calculateHash(data, previousHash) {
-  let message = data + previousHash;
-  let msgInt8 = new TextEncoder().encode(message);
-  let hashBuffer = await crypto.subtle.digest("SHA-256", msgInt8);
-  let hashArray = Array.from(new Uint8Array(hashBuffer));
-  let hashHex = hashArray.map((b) => b.toString(16).padStart(2, "0")).join("");
-  return hashHex;
-} */
